+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+- allow executing multiple tasks (add `-m`/`--multiple-tasks`)
+
 ## [0.1.24] - 2024-09-13
 
 ### Added